i965: Fix signedness of backend_reg::reg_offset.
And make it 16-bit so it packs nicely with the previous field.

+++ src/mesa/drivers/dri/i965/brw_shader.h
@@ -77,7 +77,7 @@ struct backend_reg
     *
     * For uniforms, this is in units of 1 float.
     */
-   int reg_offset;
+   uint16_t reg_offset;
 
    struct brw_reg fixed_hw_reg;
